+++ libnavigation-base/src/test/java/com/mapbox/navigation/base/route/NavigationRouteTest.kt
@@ -38,6 +38,7 @@ class NavigationRouteTest {
                 repeat(routesCount) {
                     add(
                         mockk {
+                            every { routeInfo } returns mockk(relaxed = true)
                             every { routeId } returns "$it"
                             every { routerOrigin } returns com.mapbox.navigator.RouterOrigin.ONBOARD
                         }
@@ -256,6 +257,7 @@ class NavigationRouteTest {
                 repeat(routesCount) {
                     add(
                         mockk {
+                            every { routeInfo } returns mockk(relaxed = true)
                             every { routeId } returns "some_id"
                             every { routerOrigin } returns com.mapbox.navigator.RouterOrigin.ONBOARD
                         }